Move SharedBody, ReaderOptions, and FetchResult to gosub_net
