+++ crates/gosub_net/src/net/mod.rs
@@ -2,5 +2,6 @@ pub mod events;
 pub mod null_emitter;
 pub mod observer;
 pub mod request_ref;
+pub mod shared_body;
 pub mod types;
 pub mod utils;
